split rom hashing out of hash.c into hash_rom.c
