+++ test/rcheevos-test.vcxproj
@@ -160,6 +160,7 @@
     <ClCompile Include="..\src\rhash\cdreader.c" />
     <ClCompile Include="..\src\rhash\hash.c" />
     <ClCompile Include="..\src\rhash\hash_encrypted.c" />
+    <ClCompile Include="..\src\rhash\hash_rom.c" />
     <ClCompile Include="..\src\rhash\hash_zip.c" />
     <ClCompile Include="..\src\rhash\md5.c" />
     <ClCompile Include="..\src\rurl\url.c" />
@@ -186,6 +187,7 @@
     <ClCompile Include="rhash\mock_filereader.c" />
     <ClCompile Include="rhash\test_cdreader.c" />
     <ClCompile Include="rhash\test_hash.c" />
+    <ClCompile Include="rhash\test_hash_rom.c" />
     <ClCompile Include="rhash\test_hash_zip.c" />
     <ClCompile Include="rurl\test_url.c" />
     <ClCompile Include="test.c" />
