+++ test/rhash/data.h
@@ -20,19 +20,14 @@ uint8_t* generate_pcfx_bin(uint32_t binary_sectors, size_t* image_size);
 uint8_t* generate_psx_bin(const char* binary_name, uint32_t binary_size, size_t* image_size);
 uint8_t* generate_ps2_bin(const char* binary_name, uint32_t binary_size, size_t* image_size);
 
-uint8_t* generate_atari_7800_file(size_t kb, int with_header, size_t* image_size);
 uint8_t* generate_nes_file(size_t kb, int with_header, size_t* image_size);
-uint8_t* generate_fds_file(size_t sides, int with_header, size_t* image_size);
-uint8_t* generate_nds_file(size_t mb, uint32_t arm9_size, uint32_t arm7_size, size_t* image_size);
+
 uint8_t* generate_gamecube_iso(size_t mb, size_t* image_size);
 
 uint8_t* generate_iso9660_bin(uint32_t binary_sectors, const char* volume_label, size_t* image_size);
 uint8_t* generate_iso9660_file(uint8_t* image, const char* filename, const uint8_t* contents, size_t contents_size);
 
-extern uint8_t test_rom_z64[64];
-extern uint8_t test_rom_n64[64];
-extern uint8_t test_rom_v64[64];
-extern uint8_t test_rom_ndd[64];
+void fill_image(uint8_t* image, size_t size);
 
 RC_END_C_DECLS
