+++ test/test.c
@@ -40,7 +40,12 @@ extern void test_url();
 
 extern void test_cdreader();
 extern void test_hash();
+#ifndef RC_HASH_NO_ROMS
+extern void test_hash_rom();
+#endif
+#ifndef RC_HASH_NO_ZIP
 extern void test_hash_zip();
+#endif
 
 extern void test_rapi_common();
 extern void test_rapi_user();
@@ -91,6 +96,9 @@ int main(void) {
   test_rc_libretro(); /* libretro extensions require hash support */
   test_cdreader();
   test_hash();
+ #ifndef RC_HASH_NO_ROMS
+  test_hash_rom();
+ #endif
  #ifndef RC_HASH_NO_ZIP
   test_hash_zip();
  #endif
